// mono-repo/packages/gantry-app/src/app/avr/models/avr-result-enum.model.ts
export enum AvrTagsEnum {
    EachWay = 'EACH-WAY',
    EachWayTerms = 'EACHWAYTERMS',
    EventTime = 'EVENTTIME',
    EachWayFraction = 'EACHWAYFRACTION',
    Forecast = 'FORECAST',
    Tricast = 'TRICAST',
    Distance = 'DISTANCE',
    CourseName = 'COURSENAME',
    NumEachWay = 'NUMEACHWAY',
    NumEachWayDisplay = 'NUMEACHWAYDISPLAY',
    RunnerCount = 'RUNNERCOUNT',
}

export enum AvrMessageTypeEnum {
    Result = 'RESULT',
    Off = 'OFF',
    ViewerEventCard = 'VIEWEREVENTCARD',
    NoMoreBets = 'NOMOREBETS',
}

export enum AvrMarketEnum {
    Win = 'WIN',
    Trap = 'TRAP',
}

export const AvrEventTypeEnum = {
    HorseRace: '0',
    DogRace: '1',
    MotorRace: '3',
    CourseIsJump: '1',
    CourseIsNotJump: '0',
};

export const AvrEventNamesEnum = {
    Horse: 'Horse',
    JumpHorse: 'JumpHorse',
    Dog: 'Dog',
    Motor: 'Motor',
};